requirements.txt
Django==3.2.3
djangorestframework==3.12.4
dj-database-url
uwsgi
psycopg2==2.8.3
sentry-sdk==0.20.3
hiredis==1.1.0
django-bootstrap4
urllib3
chardet
requests
git+https://github.com/GhalebKhaled/django-users-plus.git@master
django-storages
boto3==1.9.33
boto
botocore
freezegun
requests-mock
shortuuid
pyyaml
image
django-redis==5.0.0
django-rq==2.4.1
rq-scheduler==0.9
colour
######
## Mongo connectors hive-metastore
pymongo[srv]
djongo

pandas==1.3.5
pyarrow